chore(hypercall): Merge mkdir.rs into mod.rs
