+++ src/hypercall/mod.rs
@@ -10,7 +10,6 @@ use std::{
 
 use fstat::{fstat, stat};
 use getdents::getdents;
-use mkdir::mkdir;
 use uhyve_interface::{
 	GuestPhysAddr,
 	v1::{self, MAX_ARGC_ENVC},
@@ -37,7 +36,6 @@ use crate::{
 
 mod fstat;
 mod getdents;
-mod mkdir;
 
 /// `addr` is the address of the hypercall parameter in the guest's memory space. `data` is the
 /// parameter that was sent to that address by the guest.
@@ -728,6 +726,53 @@ fn lseek(syslseek: &mut LseekParams, file_map: &mut UhyveFileMap) {
 		None => {
 			warn!("lseek attempted to use an unknown file descriptor");
 			-EBADF as i64
+		}
+	};
+}
+
+/// Handles a mkdir hypercall by creating a directory on the host.
+///
+/// The guest path is resolved through the file map, so directories can only be created
+/// within mapped host directories. Unmapped paths are redirected into the sandboxed
+/// temporary directory. Virtual paths are rejected.
+/// Attempts `mkdir(host_path)` on the host, mapping the outcome to a [`MkdirResult`].
+fn host_mkdir(host_path_c: &CString) -> MkdirResult {
+	// SAFETY: `host_path_c` is a valid, null-terminated C string.
+	if unsafe { libc::mkdir(host_path_c.as_ptr(), 0o777) } < 0 {
+		MkdirResult::Error(translate_last_errno().unwrap_or(EIO))
+	} else {
+		MkdirResult::Success
+	}
+}
+
+/// Handles a mkdir hypercall by creating a directory on the host.
+///
+/// The guest path is resolved through the file map, so directories can only be created
+/// within mapped host directories. Unmapped paths are redirected into the sandboxed
+/// temporary directory. Virtual paths are rejected.
+pub(crate) fn mkdir(mem: &MmapMemory, sysmkdir: &mut MkdirParams, file_map: &mut UhyveFileMap) {
+	let Some(guest_path) = (unsafe { decode_guest_path(mem, sysmkdir.path) }) else {
+		error!("The kernel requested to mkdir() a non-UTF8 path: Rejecting...");
+		sysmkdir.ret = MkdirResult::Error(EINVAL);
+		return;
+	};
+
+	sysmkdir.ret = match file_map.get_host_path(guest_path, false) {
+		Some(UhyveMapLeaf::OnHost(host_path)) => {
+			// We can safely unwrap, as a resolved host path never contains internal NUL bytes.
+			let host_path_c = CString::new(host_path.as_os_str().as_bytes()).unwrap();
+			host_mkdir(&host_path_c)
+		}
+		Some(UhyveMapLeaf::Virtual(_)) => {
+			debug!("mkdir {guest_path:?}: target is a read-only virtual file, rejecting...");
+			MkdirResult::Error(EROFS)
+		}
+		None => {
+			debug!("mkdir {guest_path:?}: not mapped, creating a temporary directory...");
+			match file_map.create_temporary_directory(guest_path) {
+				Some(host_path_c) => host_mkdir(&host_path_c),
+				None => MkdirResult::Error(EINVAL),
+			}
 		}
 	};
 }
